Refactor accounts module to use AccountEntity instead of Account in interfaces and services
