+++ src/user-accounts-actions/controller/account-actions.controller.ts
@@ -1,12 +1,31 @@
-import { Body, Controller, Delete, UseGuards, Request } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  UseGuards,
+  Request,
+  Post,
+} from '@nestjs/common';
 import { JwtAuthGuard } from '../../auth/guards/jwt-auth.guard';
-import { DeleteAccountDto } from '../../accounts/dtos/accounts.dto';
+import {
+  CreateAccountDto,
+  DeleteAccountDto,
+} from '../../accounts/dtos/accounts.dto';
 import { AccountsActionsService } from '../services/Accounts/accounts.-actions.service';
 
 @UseGuards(JwtAuthGuard)
-@Controller('user-accounts-actions')
+@Controller('accounts-actions')
 export class UserAccountsActionsController {
   constructor(private userAccountActionsService: AccountsActionsService) {}
+
+  @Post('create-account')
+  createAccount(@Body() payload: CreateAccountDto, @Request() req) {
+    const userId = req.user.sub;
+    return this.userAccountActionsService.createAccount({
+      data: payload,
+      userId,
+    });
+  }
 
   @Delete('delete-account')
   remove(@Body() payload: DeleteAccountDto, @Request() req) {
