+++ src/user-accounts-actions/user-actions.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { UserAccountsActionsService } from './services/user-accounts-actions.service';
-import { UserAccountsActionsController } from './controller/user-accounts-actions.controller';
+import { UserAccountsActionsController } from './controller/account-actions.controller';
 import { AccountsModule } from 'src/accounts/accounts.module';
 import { RecordsModule } from 'src/records/records.module';
 import { BudgetsModule } from 'src/budgets/budgets.module';
